% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ficha_tecnica.R
\name{add_footer}
\alias{add_footer}
\title{add_footer}
\usage{
add_footer(html, footer)
}
\arguments{
\item{html}{html string}

\item{footer}{footer text}
}
\value{
html string
}
\description{
add_footer
}
